+++ helium_iceberg/src/stream/mod.rs
@@ -208,201 +208,3 @@ mod event_tests {
         assert_eq!(state.latest_sequence_number().await.unwrap(), None);
     }
 }
-
-#[cfg(all(test, feature = "test-harness"))]
-mod tests {
-    use super::*;
-    use crate::test_harness::IcebergTestHarness;
-    use crate::{FieldDefinition, PartitionDefinition, TableDefinition};
-    use iceberg::table::Table;
-    use serde::{Deserialize, Serialize};
-    use std::time::Duration;
-
-    #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-    struct Number {
-        name: String,
-        value: i64,
-    }
-
-    fn numbers_table() -> TableDefinition {
-        TableDefinition::builder("default", "numbers")
-            .with_fields([
-                FieldDefinition::required_string("name"),
-                FieldDefinition::required_long("value"),
-            ])
-            .with_partition(PartitionDefinition::identity("name"))
-            .build()
-            .expect("table def")
-    }
-
-    fn num(name: &str, value: i64) -> Number {
-        Number {
-            name: name.to_string(),
-            value,
-        }
-    }
-
-    async fn read_snapshot(table: &Table, snapshot_id: i64) -> Vec<Number> {
-        let mut rows = added_record_batches(table, snapshot_id)
-            .await
-            .unwrap()
-            .iter()
-            .flat_map(|b| batch_to_records::<Number>(b).unwrap())
-            .collect::<Vec<_>>();
-        rows.sort_by_key(|n| n.value);
-        rows
-    }
-
-    /// Snapshot ids ordered by sequence number (commit order).
-    async fn ordered_snapshot_ids(harness: &IcebergTestHarness) -> Vec<i64> {
-        let table = harness
-            .iceberg_catalog()
-            .load_table("default", "numbers")
-            .await
-            .expect("load table");
-        let mut snaps: Vec<(i64, i64)> = table
-            .metadata()
-            .snapshots()
-            .map(|s| (s.sequence_number(), s.snapshot_id()))
-            .collect();
-        snaps.sort_by_key(|(seq, _)| *seq);
-        snaps.into_iter().map(|(_, id)| id).collect()
-    }
-
-    #[tokio::test]
-    async fn reader_isolates_rows_added_by_each_snapshot() -> anyhow::Result<()> {
-        let harness = IcebergTestHarness::new_with_tables([numbers_table()]).await?;
-        let writer = harness.get_table_writer::<Number>("numbers").await?;
-
-        writer.write(vec![num("a", 1)]).await?;
-        writer.write(vec![num("b", 2), num("c", 3)]).await?;
-        writer.write(vec![num("d", 4)]).await?;
-
-        let ids = ordered_snapshot_ids(&harness).await;
-        assert_eq!(ids.len(), 3);
-
-        let table = harness
-            .iceberg_catalog()
-            .load_table("default", "numbers")
-            .await?;
-
-        assert_eq!(read_snapshot(&table, ids[0]).await, vec![num("a", 1)]);
-        assert_eq!(
-            read_snapshot(&table, ids[1]).await,
-            vec![num("b", 2), num("c", 3)]
-        );
-        assert_eq!(read_snapshot(&table, ids[2]).await, vec![num("d", 4)]);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn poller_streams_each_snapshot_in_order_and_advances_watermark() -> anyhow::Result<()> {
-        let harness = IcebergTestHarness::new_with_tables([numbers_table()]).await?;
-        let writer = harness.get_table_writer::<Number>("numbers").await?;
-        writer.write(vec![num("a", 1)]).await?;
-        writer.write(vec![num("b", 2)]).await?;
-        writer.write(vec![num("c", 3)]).await?;
-
-        // Let the poller open its own db under this dir.
-        let dir = tempfile::tempdir()?;
-        let (mut rx, server) = continuous::<Number>()
-            .catalog(harness.iceberg_catalog().clone())
-            .namespace("default")
-            .table("numbers")
-            .db_dir(dir.path())
-            .poll_duration(Duration::from_millis(200))
-            .create()
-            .await?;
-
-        let (trigger, listener) = triggered::trigger();
-        let handle = tokio::spawn(server.run(listener));
-
-        let mut seen = Vec::new();
-        let mut last_seq = i64::MIN;
-        for _ in 0..3 {
-            let mut event = rx.recv().await.expect("event");
-            // Snapshots arrive in strictly increasing sequence-number order.
-            assert!(event.snapshot.sequence_number > last_seq);
-            last_seq = event.snapshot.sequence_number;
-
-            seen.extend(event.take_rows());
-            event.commit().await?;
-        }
-
-        seen.sort_by_key(|n| n.value);
-        assert_eq!(seen, vec![num("a", 1), num("b", 2), num("c", 3)]);
-
-        trigger.trigger();
-        let _ = handle.await?;
-
-        // Watermark advanced to the latest snapshot's sequence number; reopen
-        // the poller's db (`default-numbers-default.db`) to read it back.
-        let store = StreamState::open(
-            dir.path().join("default-numbers-default.db"),
-            "default",
-            "numbers",
-        )
-        .await?;
-        assert_eq!(store.latest_sequence_number().await?, Some(last_seq));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn poller_resumes_after_watermark() -> anyhow::Result<()> {
-        let harness = IcebergTestHarness::new_with_tables([numbers_table()]).await?;
-        let writer = harness.get_table_writer::<Number>("numbers").await?;
-        writer.write(vec![num("a", 1)]).await?;
-        writer.write(vec![num("b", 2)]).await?;
-
-        // Pre-seed the watermark past the first two snapshots.
-        let ids = ordered_snapshot_ids(&harness).await;
-        let table = harness
-            .iceberg_catalog()
-            .load_table("default", "numbers")
-            .await?;
-        let second = table.metadata().snapshot_by_id(ids[1]).unwrap();
-        let second_meta = SnapshotMeta {
-            snapshot_id: ids[1],
-            sequence_number: second.sequence_number(),
-            timestamp: second.timestamp()?,
-            operation: second.summary().operation.clone(),
-        };
-
-        // Seed the watermark in the db the poller will open, then close it.
-        let dir = tempfile::tempdir()?;
-        let store = StreamState::open(
-            dir.path().join("default-numbers-default.db"),
-            "default",
-            "numbers",
-        )
-        .await?;
-        store.record(&second_meta).await?;
-        drop(store);
-
-        // A third snapshot committed after the seeded watermark.
-        writer.write(vec![num("c", 3)]).await?;
-
-        let (mut rx, server) = continuous::<Number>()
-            .catalog(harness.iceberg_catalog().clone())
-            .namespace("default")
-            .table("numbers")
-            .db_dir(dir.path())
-            .poll_duration(Duration::from_millis(200))
-            .create()
-            .await?;
-
-        let (trigger, listener) = triggered::trigger();
-        let handle = tokio::spawn(server.run(listener));
-
-        let mut event = rx.recv().await.expect("event");
-        let seen = event.take_rows();
-        event.commit().await?;
-        // Only the post-watermark snapshot is replayed.
-        assert_eq!(seen, vec![num("c", 3)]);
-
-        trigger.trigger();
-        let _ = handle.await?;
-        Ok(())
-    }
-}
